refactor(graphql): clean up, fix typos/misnomers
