+++ packages/graphql/lib/common.js
